ReponseEntity add

+++ src/main/java/gusi/springframework/projetofinancas/controlllers/ContasController.java
@@ -1,8 +1,13 @@
 package gusi.springframework.projetofinancas.controlllers;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,47 +29,77 @@ public class ContasController {
 
 	@Autowired
 	private ContasRepository contasRepository;
-	
+
 	@PostMapping
 	@Transactional
-	public void cadastrarConta(@RequestBody Conta conta) { 
-		contasRepository.save(conta);				
-				
-	}		
-	
+	public ResponseEntity<ContaDTO> cadastrarConta(@RequestBody @Valid Conta conta) {
+
+		if (ContaDTO.class.getDeclaredFields().toString().isEmpty()) {
+			return ResponseEntity.badRequest().build();
+		} else {
+			contasRepository.save(conta);
+			return ResponseEntity.ok(new ContaDTO(conta.getId(), conta.getNome(), conta.getValor(),
+					conta.getData(), conta.getCategorias()));
+		}
+	}
+
 	@GetMapping("/{id}")
-	public List<ContaDTO> listarContas(@PathVariable Long id){
-		List<Conta> contas = contasRepository.findByUsuarioId(id);
-		return ContaDTO.converter(contas);
+	public ResponseEntity<List<ContaDTO>> listarContas(@PathVariable Long id) {
+
+		if (contasRepository.findByUsuarioId(id).isEmpty()) {
+			return ResponseEntity.notFound().build();
+		} else {
+			List<Conta> contas = contasRepository.findByUsuarioId(id);
+			List<ContaDTO> result = ContaDTO.converter(contas);
+			return ResponseEntity.ok().body(result);
+		}
 	}
-	
+
 	@GetMapping("/categorias/{id}/{ano}")
-	public List<SomaCategoria> listarContasCategoria(@PathVariable Long id, @PathVariable String ano) {		
-		List<SomaCategoria> contas = contasRepository.somarCategorias(id, ano);
-		return contas;
-		
-	}	
-	
-	
+	public ResponseEntity<SomaCategoria> listarContasCategoria(@PathVariable Long id, @PathVariable String ano) {
+
+		if (id == null | ano == null) {
+			return ResponseEntity.notFound().build();
+		} else {
+			Optional<SomaCategoria> contas = contasRepository.somarCategorias(id, ano);
+			return ResponseEntity.ok(contas.get());
+		}
+
+	}
+
 	@GetMapping("/{id}/{mes}/{ano}")
-	public List<ContaDTO> listarContas(@PathVariable Long id, @PathVariable String mes, @PathVariable String ano){
-		List<Conta> contas = contasRepository.listarContasMesAno(id,mes, ano);
-		return ContaDTO.converter(contas);
+	public ResponseEntity<List<ContaDTO>> listarContas(@PathVariable Long id, @PathVariable String mes,
+			@PathVariable String ano) {
+
+		if (id == null | mes == null | ano == null) {
+			return ResponseEntity.notFound().build();
+		} else {
+			List<Conta> contas = contasRepository.listarContasMesAno(id, mes, ano);
+			List<ContaDTO> result = ContaDTO.converter(contas);
+			return ResponseEntity.ok().body(result);
+		}
 	}
-	
+
 	@GetMapping("/totalmensal/{id}/{mes}/{ano}")
-	public List<TotalMensal> listarTotalMensal(@PathVariable Long id, @PathVariable String mes, @PathVariable String ano) { 
-		List<TotalMensal> contas = contasRepository.listarTotalMensal(id,mes,ano);
-		return contas;
+	public ResponseEntity<List<TotalMensal>> listarTotalMensal(@PathVariable Long id, @PathVariable String mes,
+			@PathVariable String ano) {
+		if (id == null | mes == null | ano == null) {
+			return ResponseEntity.notFound().build();
+		} else {
+			List<TotalMensal> contas = contasRepository.listarTotalMensal(id, mes, ano);
+			return ResponseEntity.ok().body(contas);
+		}
 	}
-	
-	
-	
+
 	@DeleteMapping("/{id}")
 	@Transactional
-	public void deletarConta(@PathVariable Long id) {
-		contasRepository.deleteById(id);		
+	public ResponseEntity<?> deletarConta(@PathVariable Long id) {
+		if (id == null) {
+			return ResponseEntity.notFound().build();
+		} else {
+			contasRepository.deleteById(id);
+			return new ResponseEntity<>("Conta Deletada", HttpStatus.OK);
+		}
 	}
-	
-	
+
 }
